+++ client/src/App.jsx
@@ -31,13 +31,15 @@ const client = new ApolloClient({
 
 function App() {
   return (
-    
+    <ApolloProvider client={client}>
       <div>
         {/* <StoreProvider> */}
           <Nav />
           <Outlet />
         {/* </StoreProvider> */}
+        
       </div>
+      </ApolloProvider>
  
   );
 }
