+++ server/schemas/resolvers.js
@@ -50,7 +50,7 @@ const resolvers = {
         const serviceObjects = [];
         for (const serviceData of services) {
           const { _id, addOns } = serviceData;
-          console.log(_id)
+          
           const service = await Services.findById(_id);
           if (!service) {
             throw new Error(`Service not found`);
@@ -60,7 +60,7 @@ const resolvers = {
           }
           serviceObjects.push(service._id);
         } 
-        console.log(serviceObjects)
+       
         const newBooking = await Booking.create({
           user: userId,
           service: serviceObjects,
@@ -68,7 +68,7 @@ const resolvers = {
           date,
           time
         });
-console.log( await newBooking.populate("service"))
+
         return await newBooking.populate("service");
       } catch (error) {
         console.log(error)
